[ML] Fix Create a data view option not working in Transforms (#168263)
(cherry picked from commit 6213e41b6cc4915e9003999949d7c190e05ac256)
